% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_knockoff.R
\name{DUET_knockoff}
\alias{DUET_knockoff}
\title{FDR-Controlled Joint Feature Selection from Dual Microbiome data.}
\usage{
DUET_knockoff(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test"
)
}
\arguments{
\item{W}{Count data.}

\item{class_K}{Vector data, representing different sources of count data.}

\item{data_x}{Covariate matrix.}

\item{M}{The sequencing depth of count data.}

\item{y}{Response variable, typically binary.}

\item{T_var}{True correlation variables, default to NULL.}

\item{fdr}{The target FDR level, default 0.2.}

\item{offset}{value between 0 and 1.}

\item{test_statistic}{Type of single-platform test statistic, default 'DE', c('DE', 'GLM', 'RF')}

\item{filter_statistics}{Cross-platform aggregated knockoff contrast used to rank features:
1 = product rule; 2 = max rule; 3 = sum rule.}

\item{test1}{Test method, select when test_statistic = 'DE', default is 'wilcox.test', c('wilcox.test', 'ks.test').}
}
\value{
A list with:
  \item{test_stat}{Matrix of source‐specific test statistics.}
  \item{filter_stat}{Aggregated cross-platform statistic.}
  \item{S}{Selected feature indices.}
  \item{res}{Named vector of target FDR, empirical FDR, and power.}
}
\description{
FDR-Controlled Joint Feature Selection from Dual Microbiome data.
}
\examples{
library(DUETknockoff)

set.seed(42)

n_group <- 2
n_data <- 2

W1 <- matrix(rpois(3000, lambda=5), nrow=30, ncol=100)
W2 <- matrix(rpois(3000, lambda=50), nrow=30, ncol=100)
W <- rbind(W1, W2)
M <- rowSums(W)
data_x <- NULL

n_sam <- nrow(W1)
class_K <- factor(rep(1:n_data, each = n_sam))
y <- rep(c(rep(1, n_sam/2), rep(2, n_sam/2)), times = n_data)

res.DUET.knockoff <- DUET_knockoff(W = W, M = M, class_K = class_K, y=y, data_x=NULL,
                     T_var = NULL, test_statistic = "DE",
                     filter_statistics=3, offset=1)
res.DUET.knockoff$test_stat[1:10]
res.DUET.knockoff$filter_stat[1:10]
}
