% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUET_knockoff_others.R
\name{stat.glmnet_coefdiff}
\alias{stat.glmnet_coefdiff}
\title{Importance statistics based on GLM with cross-validation.}
\usage{
stat.glmnet_coefdiff(X, X_k, y, family = "gaussian", cores = 2, ...)
}
\arguments{
\item{X}{Numeric matrix (n × p). Original variables.}

\item{X_k}{Numeric matrix (n × p). Knockoff variables.}

\item{y}{Response vector of length n. Accepts numeric (gaussian, poisson),
factor/integer (binomial, multinomial), or survival matrix for \code{family="cox"}.}

\item{family}{response type, \code{"gaussian"}.}

\item{cores}{Number of cores used for parallel CV, default 2.}

\item{...}{additional arguments specific to \code{glmnet}.}
}
\value{
A vector of statistics \eqn{W} of length p.
}
\description{
Computes importance statistics
\eqn{W_j = |Z_j| - |\tilde Z_j|} using cross‐validated
penalized generalized linear models from **glmnet**.
}
\details{
This function fits a penalized GLM using \code{glmnet} with 10-fold
cross-validation, extracts the fitted coefficients of
\code{(X, X_k)}, and constructs knockoff statistics by
comparing the magnitudes of corresponding coefficients.
}
\examples{
set.seed(2022)
p=200; n=100; k=15
mu = rep(0,p); Sigma = diag(p)
X = matrix(rnorm(n*p),n)
nonzero = sample(p, k)
beta = 3.5 * (1:p \%in\% nonzero)
y = X \%*\% beta + rnorm(n)
knockoffs = function(X) create.gaussian(X, mu, Sigma)

# Basic usage with default arguments
result = knockoff.filter(X, y, knockoffs=knockoffs,
                           statistic=stat.glmnet_coefdiff)
print(result$selected)

# Advanced usage with custom arguments
foo = stat.glmnet_coefdiff
k_stat = function(X, X_k, y) foo(X, X_k, y, nlambda=200)
result = knockoff.filter(X, y, knockoffs=knockoffs, statistic=k_stat)
print(result$selected)

}
\concept{statistics}
