% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ye_count1.R
\docType{data}
\name{Ye_count1}
\alias{Ye_count1}
\title{Simulated OTU count table}
\format{
A data frame with 44 observations on 304 variables.
}
\usage{
data("Ye_count1")
}
\description{
A processed OTU count table contains 44 subjects with 304 OTUs.
}
\examples{
data(Ye_count1)
}
\keyword{datasets}
