% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zhu_count1.R
\docType{data}
\name{Zhu_count1}
\alias{Zhu_count1}
\title{Simulated OTU count table}
\format{
A data frame with 81 observations on 402 variables.
}
\usage{
data("Zhu_count1")
}
\description{
A processed OTU count table contains 44 subjects with 304 OTUs.
}
\examples{
data(Zhu_count1)
}
\keyword{datasets}
